+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,6 +1,7 @@
 2.13.3 (unreleased)
 --------------------
 * Added function to retrieve local roles for a specified user [bogdatan]
+* fix FileIterator interface for zip download to work [simiamih]
 
 2.13.2 (2012-04-05)
 --------------------